Safe Loader: move docstring related functions to the safeloader module
