+++ avocado/core/loader.py
@@ -30,6 +30,7 @@ import sys
 from . import data_dir
 from . import output
 from . import test
+from . import safeloader
 from ..utils import path
 from ..utils import stacktrace
 from .settings import settings
@@ -37,28 +38,6 @@ from .settings import settings
 DEFAULT = False  # Show default tests (for execution)
 AVAILABLE = None  # Available tests (for listing purposes)
 ALL = True  # All tests (including broken ones)
-
-
-#: Gets the tag value from a string. Used to tag a test class in various ways
-AVOCADO_DOCSTRING_TAG_RE = re.compile(r'\s*:avocado:\s*(\S+)\s*')
-
-
-def get_docstring_tag(docstring):
-    if docstring is None:
-        return None
-    result = AVOCADO_DOCSTRING_TAG_RE.search(docstring)
-    if result is not None:
-        return result.groups()[0]
-
-
-def is_docstring_tag_enable(docstring):
-    result = get_docstring_tag(docstring)
-    return result == 'enable'
-
-
-def is_docstring_tag_disable(docstring):
-    result = get_docstring_tag(docstring)
-    return result == 'disable'
 
 
 class LoaderError(Exception):
@@ -593,9 +572,9 @@ class FileLoader(TestLoader):
                 docstring = ast.get_docstring(statement)
                 # Looking for a class that has in the docstring either
                 # ":avocado: enable" or ":avocado: disable
-                if is_docstring_tag_disable(docstring):
+                if safeloader.is_docstring_tag_disable(docstring):
                     continue
-                elif is_docstring_tag_enable(docstring):
+                elif safeloader.is_docstring_tag_enable(docstring):
                     functions = [st.name for st in statement.body if
                                  isinstance(st, ast.FunctionDef) and
                                  st.name.startswith('test')]
