+++ avocado/core/safeloader.py
@@ -17,6 +17,7 @@ Safe (AST based) test loader module utilities
 """
 
 import ast
+import re
 
 
 def modules_imported_as(module):
@@ -58,3 +59,41 @@ def modules_imported_as(module):
                 else:
                     result[name.name] = name.name
     return result
+
+
+#: Gets the tag value from a string. Used to tag a test class in various ways
+AVOCADO_DOCSTRING_TAG_RE = re.compile(r'\s*:avocado:\s*(\S+)\s*')
+
+
+def get_docstring_tag(docstring):
+    """
+    Returns the value of the avocado custom tag inside a docstring
+
+    :param docstring: the complete text used as documentation
+    :type docstring: str
+    """
+    if docstring is None:
+        return None
+    result = AVOCADO_DOCSTRING_TAG_RE.search(docstring)
+    if result is not None:
+        return result.groups()[0]
+
+
+def is_docstring_tag_enable(docstring):
+    """
+    Checks if there's an avocado tag that enables its class as a Test class
+
+    :rtype: bool
+    """
+    result = get_docstring_tag(docstring)
+    return result == 'enable'
+
+
+def is_docstring_tag_disable(docstring):
+    """
+    Checks if there's an avocado tag that disables its class as a Test class
+
+    :rtype: bool
+    """
+    result = get_docstring_tag(docstring)
+    return result == 'disable'
